requirements.txt
# Microfluidic Property Prediction GUI - Python Dependencies
# Tested on Python 3.10/3.11 (recommended; if you use 3.12+, drop pyopengltk)

# Deep Learning Framework
torch==2.2.2
torchvision==0.17.2

# Image Processing
Pillow==10.2.0
numpy==1.26.4

# Web API
fastapi==0.110.1
uvicorn[standard]==0.30.3
httpx==0.27.2
pytest==8.3.3

# 3D Visualization and Mesh Processing
vispy==0.13.0
trimesh==4.4.0

# Geometric Operations
shapely==2.0.4

# Data Processing and Machine Learning
joblib==1.3.2
scikit-learn==1.5.2

# Plotting and Visualization
matplotlib==3.8.4

# Image/mesh utilities (required for trimesh marching cubes)
scikit-image==0.22.0

# GUI Backend for Vispy (Tkinter)
pyopengltk==0.0.4 ; python_version < "3.12"
PyOpenGL==3.1.7

# Optional: For better performance (Python <=3.11)
# numba==0.59.1  # Uncomment if needed for performance optimization
